refactor: 각 도메인 별 레이어 나눔
