+++ src/main/java/com/danawa/dsearch/server/collections/service/CollectionService.java
@@ -7,8 +7,8 @@ import com.danawa.dsearch.server.config.ElasticsearchFactory;
 import com.danawa.dsearch.server.collections.entity.Collection;
 import com.danawa.dsearch.server.excpetions.DuplicateException;
 import com.danawa.dsearch.server.excpetions.IndexingJobFailureException;
-import com.danawa.dsearch.server.clusters.ClusterService;
-import com.danawa.dsearch.server.indices.IndicesService;
+import com.danawa.dsearch.server.clusters.service.ClusterService;
+import com.danawa.dsearch.server.indices.service.IndicesService;
 import com.danawa.dsearch.server.utils.JsonUtils;
 import com.google.gson.Gson;
 import org.apache.http.util.EntityUtils;
