+++ src/test/java/com/danawa/dsearch/server/YmlDictTest.java
@@ -1,6 +1,6 @@
 package com.danawa.dsearch.server;
 
-import com.danawa.dsearch.server.dictionary.DictionaryService;
+import com.danawa.dsearch.server.dictionary.service.DictionaryService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
